feat: Add GPU rendering to canvas.
This change adds WebGPU rendering to the canvas element. Underlying GPU compute context, pipeline, vertex buffers, command encoder were created, adjusted, and used to render triangle to the screen. The function that generates all this set was added as an effect to be called on component mount and on each change of the GPU and ctx states.